tests do not pass because no product functionality yet. i am sed.

+++ code/src/main/java/com/sadna/sadnamarket/domain/discountPolicies/DiscountPolicy.java
@@ -68,6 +68,6 @@ public abstract class DiscountPolicy {
         this.productsId = productsId;
     }
 
-    public abstract int giveDiscount(ProductDTO productDTO, CartItemDTO cartItemDTO);
+    public abstract double giveDiscount(ProductDTO productDTO, CartItemDTO cartItemDTO);
 
 }
